Extract node variation and fix hierarchy relation child node assignment
(was findIngoingHierarchyRelationsFor, now is
findIngoingHierarchyRelationsForNodeAggregate)
